Sped up like 100x mapping performance by introducing write buffer Added WriteBuffer (which is shockingly elegant) Added supplementary ByteArray encode extensions for writing Longs, Ints and Shorts

+++ src/main/kotlin/dict/spimi/SPIMIFile.kt
@@ -77,7 +77,7 @@ class SPIMIFile(file: File) : Closeable, Iterable<SPIMIEntry>, RandomAccess {
                 DocumentID(
                     flags.dicAction(
                            big = { bytedocs.decodeInt(idx) },
-                        medium = { bytedocs.decodeShort(idx) },
+                        medium = { bytedocs.decodeShort(idx).toInt() },
                          small = { bytedocs[idx].toInt() })
                 )
             }
